描画処理をmainから独立
