+++ CHAPTER6/main/__init__.py
@@ -1,5 +1,5 @@
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃処理
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-from .main  import ゲーム本体 as ゲーム開始
+from .main  import 本体 as ゲーム開始
 from .DB    import *
