+++ CHAPTER6/main/main.py
@@ -8,11 +8,12 @@ import main.DB
 from   シーン        import *
 from   背景	         import 背景作成
 from   .更新処理     import 更新処理
+from   .描画処理     import 描画処理
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃メイン
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class ゲーム本体:
+class 本体:
     #┌───────────────────────────────────
     #│初期化
     #└───────────────────────────────────
@@ -28,7 +29,7 @@ class ゲーム本体:
         main.DB.FNシーン[ シーンID.終了画面 ].FN移動.Fn切替()
         #│
         #○フレーム処理に、処理セットを登録する
-        pyxel.run(self.FN更新処理.実行, self.描画処理)
+        pyxel.run(self.FN更新処理.実行, self.FN描画処理.実行)
 		#┴
 	#────────────────────────────────────	
     def 初期化_リソース(self):
@@ -60,6 +61,7 @@ class ゲーム本体:
 		#┬
         #≫更新処理のオブジェクト・セットを作る(タプル型）
         self.FN更新処理 = 更新処理()
+        self.FN描画処理 = 描画処理()
         #│
         #≫シーン処理のオブジェクト・セットを作る(辞書型）
         main.DB.FNシーン = {
@@ -70,57 +72,3 @@ class ゲーム本体:
         #≫背景をオブジェクト化する
         main.DB.obj背景 = 背景作成()
         #┴
-
-	#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-	#┃２．描画処理
-	#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-    def 描画処理(self):
-		#┬
-        #○画面をクリアする
-        pyxel.cls(0)
-        #│
-        #●画面に各種オブジェクトを描画する
-        self.Fn描画((
-                main.DB.obj背景    ,
-                main.DB.obj敵機    ,
-                main.DB.objアイテム,
-                main.DB.obj自機    ,
-                main.DB.obj弾_自機 ,
-                main.DB.obj弾_敵機 ,
-                main.DB.obj爆発    ,
-                main.DB.FNシーン[ main.DB.シーン])) # ※ シーンに合わせて変更
-        #│
-        #〇ゲーム情報（得点／難易度）を描画する
-        pyxel.text( 5, 2, f"SCORE:{ main.DB.得点   }", 7)
-        pyxel.text(85, 2, f"LEVEL:{ main.DB.難易度 }", 7)
-        #┴
-	#────────────────────────────────────	
-    def Fn描画(self, 引数_処理一覧):
-		#┬
-		#◎└┐引数からオブジェクトを抽出し描画する
-        for tmp引数要素 in 引数_処理一覧:
-			#│＼（すべての処理を終えた場合）
-            #│ ↓
-            #│ ▼繰り返し処理を抜ける
-            #│
-            #◇┐オブジェクトを描画する
-            if isinstance(tmp引数要素, list):
-            #　├┐（引数の要素が一覧の場合）
-                #↓
-                #◎└┐更に分解し描画する
-                for tmp更に分解 in tmp引数要素:
-                    #│＼（すべての処理を終えた場合）
-                    #│ ↓
-                    #│ ▼繰り返し処理を抜ける
-                    #│
-                    #●描画処理を実行する
-                    tmp更に分解.FN描画.実行()
-                    #┴ 
-
-            elif tmp引数要素 is not None:
-            #　├┐（引数の要素がオブジェクト化している場合）
-                #↓
-                #●描画処理を実行する
-                tmp引数要素.FN描画.実行()
-            #　└┐（その他）
-        #┴　┴　┴
