+++ CHAPTER6/main/描画処理/main.py
@@ -1,0 +1,79 @@
+#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+#┃技術評論社 ゲームで学ぶPython！ CHAPTER6:MAGA WING
+#┠─────────────────────────────────────
+#┃処理（ゲーム本体）
+#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+import pyxel
+import main.DB
+from   シーン        import *
+
+#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+#┃メイン
+#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+class 本体:
+    #┌───────────────────────────────────
+    #│初期化
+    #└───────────────────────────────────
+    def __init__(self):
+		#┬
+        #≫処理セットを作る(辞書型）
+        main.DB.FNシーン = {
+            シーンID.タイトル画面   : タイトル画面(),
+            シーンID.プレイ画面     : プレイ画面()  ,
+            シーンID.終了画面       : 終了画面()    }
+        #┴
+
+	#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+	#┃処理実行
+	#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+    def 実行(self):
+		#┬
+        #○画面をクリアする
+        pyxel.cls(0)
+        #│
+        #●画面に各種オブジェクトを描画する
+        self.Fn描画((
+                main.DB.obj背景    ,
+                main.DB.obj敵機    ,
+                main.DB.objアイテム,
+                main.DB.obj自機    ,
+                main.DB.obj弾_自機 ,
+                main.DB.obj弾_敵機 ,
+                main.DB.obj爆発    ,
+                main.DB.FNシーン[ main.DB.シーン])) # ※ シーンに合わせて変更
+        #│
+        #〇ゲーム情報（得点／難易度）を描画する
+        pyxel.text( 5, 2, f"SCORE:{ main.DB.得点   }", 7)
+        pyxel.text(85, 2, f"LEVEL:{ main.DB.難易度 }", 7)
+        #┴
+
+	#────────────────────────────────────	
+    def Fn描画(self, 引数_処理一覧):
+		#┬
+		#◎└┐引数からオブジェクトを抽出し描画する
+        for tmp引数要素 in 引数_処理一覧:
+			#│＼（すべての処理を終えた場合）
+            #│ ↓
+            #│ ▼繰り返し処理を抜ける
+            #│
+            #◇┐オブジェクトを描画する
+            if isinstance(tmp引数要素, list):
+            #　├┐（引数の要素が一覧の場合）
+                #↓
+                #◎└┐更に分解し描画する
+                for tmp更に分解 in tmp引数要素:
+                    #│＼（すべての処理を終えた場合）
+                    #│ ↓
+                    #│ ▼繰り返し処理を抜ける
+                    #│
+                    #●描画処理を実行する
+                    tmp更に分解.FN描画.実行()
+                    #┴ 
+
+            elif tmp引数要素 is not None:
+            #　├┐（引数の要素がオブジェクト化している場合）
+                #↓
+                #●描画処理を実行する
+                tmp引数要素.FN描画.実行()
+            #　└┐（その他）
+        #┴　┴　┴
