+++ CHAPTER6/main/更新処理/main.py
@@ -15,7 +15,7 @@ class 本体:
     #└───────────────────────────────────
     def __init__(self):
 		#┬
-        #≫更新処理のオブジェクト・セットを作る(タプル型）
+        #≫処理セットを作る(タプル型）
         self.機能一覧 = (
             移動クラス() ,
             発射クラス() ,
@@ -24,7 +24,7 @@ class 本体:
         #┴
 
 	#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-	#┃１．更新処理
+	#┃処理実行
 	#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
     def 実行(self):
 		#┬
